update to include code to set the RTC
Added routine to change the stored the RTC on reset - hold button a
while resetting to enter this mode - see source listing for further
instructions.

I ran of of memory while adding the time set feature so have had to make
a number of other changes to reduce the code size, but functionality is
still the same as R3 version.